+++ src/main/java/module-info.java
@@ -22,7 +22,6 @@ module jobj.reflect.main
 {
 	requires lombok;
 	requires jobj.core.main;
-	requires org.objenesis;
 	requires java.logging;
 
 	exports io.github.astrapi69.reflection;