+++ CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGED:
 - update of gradle-plugin dependency 'org.ajoberstar.grgit:grgit-gradle' to new version 5.2.1
 - increase code coverage
 - moved all factory methods from class ReflectionExtensions to new factory class InstanceFactory and tagged them as deprecated
+- remove of unused dependency objenesis
 
 Version 2.1
 -------------